+++ crates/authly-common/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - namespace level added to property mapping protobuf.
 - decouple policies from services, introduce `domain` concept into document model.
 - improved implementation of policy engine with new logical model for policy combinations.
+- `entity-attribute-binding` renamed to `entity-attribute-assignment`. Now accepts only one parameter for identifying the entity.
 
 ## [0.0.5] - 2025-01-30
 ### Changed